change of migrations

+++ aspnet-core/src/Dairy.EntityFrameworkCore/Migrations/DairyDbContextModelSnapshot.cs
@@ -1516,9 +1516,9 @@ namespace Dairy.Migrations
 
             modelBuilder.Entity("Dairy.Dairy.Products.Product", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint")
+                        .HasColumnType("int")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<DateTime>("CreationTime")
